Adding Queue and ImprovedQueue classes along with unit tests.
Queue class implements a queue via a circular array and tracks
the size of the queue via the dedicated class field.
ImprovedQueue improves upon Queue by removing this field and
allocating max_queue_size+1 elements in the array. In this way
the HEAD and TAIL fields are enough to calculate the size of
the queue as well as check if the queue is full or empty.