PriorityQueue class re-implemented.
The previous implementation was unnecessary complicated by using
the circular array. This patch simplifies the implementation by
sorting the elements in the underlying array in descending order,
so that the smallest element is always at index SIZE-1 (HEAD of the
queue) and the largest element is always at index 0 (TAIL of the
queue).