Introducing QuickSortBasic class with its unit tests.
The QuickSortBasic class implements a classic QuickSort algorithm
where the pivot is always the right-most element of the current
sub-array. While simple, this is not the best choice of pivot
because the algorithm degrades to O(N^2) time complexity for
certain array layouts (for example, if the array is sorted in
reversed order).