QuickSortMedianOfThree class completed and its unit tests introduced.
The class implements the more efficient variant of the QuckSort
algorithm where the pivot is calculated as the median value of
the start, middle and end elements of the current partition.
Such pivot selection scheme significantly reduces the chance
of selecting a pivot that is a smallest or largest value in
the array, and hence reduces the possibility that algorithm
will degrade to O(N^2) performance.

Change-Id: If61c7a4f432d0122b0d450d1e4d2bc9db25b3239