Introducing RadixSortBinary class with its unit tests.
The class implements the binary RadixSort algorithm that places
the items in bucket 0 and 1 based on the bit value. For each of
the 64 bits of the key, the algorithm iterates over the array
and places items in the two buckets. The algorithm differentiates
between negative and positive numbers only once MSB bit has been
reached. In this last iteration, the negative numbers (those with
MSB bit set to 1) are placed in bucket 0 while positive numbers
(those with MSB bit set to 0) are placed in the bucket 1 (which
is the opposite of how items are placed in buckets in all the
other iterations). This makes sure that already sorted partition
of negative numbers appear before the already sorted partition
of positive numbers with zeros in between.