Introducing BinarySearchTreeViaArray class with its unit tests.
The class implements the classic binary search tree but uses an
array (actually a vector) to store the nodes instead of having
each node point to its children. The insertion, lookup and the
common traversal operations (in-order, pre-order and post-order)
have been implemented. Deletion will be implemented in the next
commit.

Change-Id: I13d2e33026f3c2de52fb401aabb6e6d079762a86