HashTableLinearProbe: underlying array used as a ring-buffer
When mapping a new key, the implementation will wrap around
if it reaches the end of the array before a free cell is
found and start looking at the start of the array. This
singnificantly improves the hash table efficiency as array
doesn't have to be resized when there's no space left at
the end of the array. Previous implementation required
the array to be resized whenever at least 2 keys map to
the very last array index.

Change-Id: Ifc735d0baf1f8f667d90e0df56d19f763de9b869