IGraph clean-up and introducing FindMinimumSpanningTree.
IGraph interface has been simplified be removing the
methods that accept the vertices as arguments and use
GetIndexOf to get their indices. If callers wishes to
use vertices instead of their indices, they can use
GetIndexOf to get the vertex index and pass that to
the graph methods.

FindMinimumSpanningTree has been introduced to the
IGraph and implemented by UndirectedUnweightedGraph.
The methods fins a minimum spanning tree given a
starting vertex. Additional unit tests have been
added to the GraphTest.

Change-Id: Iaf6f0a13c0af911ab4e3008fb5708d0e2f85ea6b