Updating the topological sort algorithm.
The topological sort algorithm has been aligned with
the DFS-based algorithm presented in 'Introduction
to algorithms, 3rd Edition'. The algorithm uses
two types of markers:

- Permanent marker used to mark a vertex once it is
added to the sorted array. If a permantently marked
vertex is reached later the DFS search terminates
there as we know that all its successor nodes have
already been processed.
- Temporary marked used to discover cycles in the
graph. The vertex is marked with this marker before
its successors are processed recursively. If the
algorithm ever reaches a vertex that is marked
with the temporary marker, the graph contains
cycles and topological sort doesn't exist.

A unit test assertint that FindTopologicalSort throws
an exception if graph is cyclic has been introduced
to the DirectedUnweightedGraphTest test fixture.

Change-Id: I53f856482f8060a19e4700ee20acade6f7c0988c