Fixing UndirectedWeightedGraphTest.TestFindingMinimumSpanningTree test
The expected MST used by the test was just one of possible
MST trees for the given graph. The mismatch came from the
fact that the priority queue used by the implementation
ordered the edges with the same weight differently than
what was expected by the test. The test has been rewritten
to indirectly verify the generated MST instead, to make
sure that the test passes for any valid MST.

Change-Id: I950cc33e6502ef6e46bdc5b8976abf83e1ba2ab7