Extending responsive_modal_gallery.html page.
Navigation functionality added to the modal window.
The user can now use left/right buttons and left/right
keyboard keys to navigate the image gallery. The user
can also exit the modal window by clicking on the
exit button or by pressing an ESCAPE key. Additionally,
when modal window is shown, the image and text that
is shown in it is animated.

Change-Id: If39a1d6b96d0b622043bb3ae81dbe293273e767c