Adding loc8r-static-prot project to webdev/mean/loc8r
Loc8r is a web application that lets users find a place with
good wifi where they can sit and get some work done.

loc8r-static-prot is a static prototype that has most of the UI
completed but where all the data is hardcoded in the views. This
is the first step of development whose aim is to get a better
understanding of how the app will look like, what data does app
need and give a chance to make bigger changes before real coding
has begun.

A Gulp-based build system is included as well.